move common methods into abstractintegrationaltest
